projects/ento_linguistics/pyproject.toml
# Ento-Linguistics Project Configuration
# Inherits workspace settings from root pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ento-linguistic-research"
version = "1.0.0"
description = "Ento-Linguistic analysis of scientific terminology in entomology"
authors = [
    {name = "Daniel Ari Friedman", email = "daniel@activeinference.institute"}
]
requires-python = ">=3.10"
dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    # Natural language processing
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "scikit-learn>=1.3.0",
    # Network analysis and visualization
    "networkx>=3.2.0",
    "plotly>=5.17.0",
    "seaborn>=0.13.0",
    # System monitoring (required by infrastructure)
    "psutil>=5.9.0",
    # Infrastructure dependencies
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-httpserver>=1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow-running (deselect with '-m not slow')",
    "ollama: marks tests requiring a live Ollama service",
    "tts: marks tests requiring text-to-speech dependencies",
    "integration: marks full end-to-end integration tests",
]

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = ["tests/*", "*/__init__.py", "*/test_*.py"]


[tool.coverage.report]
fail_under = 90
show_missing = true
precision = 2
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.uv]
# Inherits workspace settings from root pyproject.toml
